mac: add support for HMAC with GOST hash algorithms
The following three message digest algorithms have been added along
with corresponding HMAC variants:

*  gosthash    -- GOST R 34.11-94, 256 bit;
*  stribog-256 -- GOST R 34.11-2012, 256 bit;
*  stribog-512 -- GOST R 34.11-2012, 512 bit.

HMAC with GOST R 34.11-94 aka Gosthash is not enabled by default, since
the latter is officially considered obsolete. However, it is still widely
used and has no known vulnerabilities. Moreover, it uses internally the
GOST 28147-89 cipher algorithm, which is adopted as part of the new GOST
R 34.12-2015 standard under the name Magma (256 bit key size, 64 bit
block size), which also has no known vulnerabilities.

The main reason for exclusion Gosthash from the new standard is its
significantly lower operating speed compared to the new Stribog algorithm.

The Stribog algorithm with a digest size of 256 bits is, in fact, the
same Stribog algorithm with a digest size of 512 bits with a different
IV and a truncated lower half of the output value and nothing more.