Add support for algorithms provided by an engine
Some algorithms may be provided by the OpenSSL engine, so they may not
be available in the current configuration. Moreover, algorithms can be
implemented in hardware that may be temporarily unavailable. Let's just
ignore the missing algorithms.

This change does not affect the algorithms, the implementation of which
is built into OpenSSH and the crypto-library used.