cipher: add support for GOST cipher algorithms
The following two groups of cipher algorithms have been added:

*  GOST 28147-89, GOST R 34.12-2015 256-bit key, 64-bit block:

   *  gost89-cbc -- CBC mode;
   *  gost89-cfb -- CFB mode;
   *  gost89-ofb -- OFB mode;
   *  gost89-ctr -- CTR mode.

*  GOST R 34.12-2015 256-bit key, 128-bit block aka Kuznechik:

   *  kuznechik-cbc -- CBC mode;
   *  kuznechik-cfb -- CFB mode;
   *  kuznechik-ctr -- CTR mode;
   *  kuznechik-ofb -- OFB mode.

Note that the key meshing mode is incompatible with the current procedure
for moving the state after authentication: to restore the state, it is not
enough to know the initial key and the current initialization vector. We
disable key and IV meshing mode where it is enabled by default.

The name of GOST R 34.12-2015 128-bit block cipher has been transliterated
from Cyrillic to Latin according to the ISO/IEC 7501-1-2013 standard in
force in Russia: do not try to replace it with "Kuznyechik" -- that's
a mistake in standard.